removed unnecessary interfaces
