--- api/service/src/main/kotlin/wasik/api/service/mapper/item/CommonItemDataMapperImpl.kt
@@ -1,15 +1,0 @@
-package wasik.api.service.mapper.item
-
-import kotlinx.coroutines.async
-import kotlinx.coroutines.coroutineScope
-import domain.model.item.CommonItemData
-import org.springframework.stereotype.Component
-import wasik.api.model.IItem
-
-@Component
-class CommonItemDataMapperImpl(private val rarityMapper: RarityMapper) : CommonItemDataMapper {
-    override suspend fun mapToCommonItemInfo(item: IItem): CommonItemData = coroutineScope {
-        val rarity = async { rarityMapper.mapToRarity(item.rarity) }
-        CommonItemData(item.name, rarity.await(), item.value, item.weight, item.description)
-    }
-}
