+++ api/service/src/main/kotlin/wasik/api/service/mapper/item/PropertyMapper.kt
@@ -1,8 +1,12 @@
 package wasik.api.service.mapper.item
 
+import org.springframework.stereotype.Component
 import wasik.api.model.Property
 import domain.model.misc.Property as DomainProperty
 
-interface PropertyMapper {
-    suspend fun mapToProperty(property: Property): DomainProperty
+@Component
+class PropertyMapper {
+    suspend fun mapToProperty(property: Property): DomainProperty {
+        return DomainProperty(property.name, property.description)
+    }
 }
