+++ api/service/src/main/kotlin/wasik/api/service/mapper/item/damage/DamageMapper.kt
@@ -1,7 +1,19 @@
 package wasik.api.service.mapper.item.damage
 
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
+import org.springframework.stereotype.Component
 import wasik.api.model.model.damage.Damage
+import domain.model.damage.Damage as DomainDamage
 
-interface DamageMapper {
-    suspend fun mapToDamage(damage: Damage) : domain.model.damage.Damage
+@Component
+class DamageMapper(
+    private val dieMapper: wasik.api.service.mapper.item.DieMapper,
+    private val damageTypeMapper: DamageTypeMapper
+) {
+    suspend fun mapToDamage(damage: Damage): DomainDamage = coroutineScope {
+        val die = async { dieMapper.mapDie(damage.dieType) }
+        val damageType = async { damageTypeMapper.mapToDamageType(damage.damageType) }
+        DomainDamage(damageType.await(), damage.dieAmount, die.await(), damage.bonus)
+    }
 }
