--- api/service/src/main/kotlin/wasik/api/service/mapper/item/weapon/TypeMapperImpl.kt
@@ -1,17 +1,0 @@
-package wasik.api.service.mapper.item.weapon
-
-import org.springframework.stereotype.Component
-import wasik.api.model.model.weapon.WeaponType
-import domain.model.item.weapon.WeaponType as DomainWeaponType
-
-@Component
-class TypeMapperImpl : TypeMapper {
-    override suspend fun mapToType(weaponType: WeaponType): DomainWeaponType {
-        return when(weaponType) {
-            WeaponType.MELEE, WeaponType.TWO_HANDED_MELEE -> DomainWeaponType.MELEE
-            WeaponType.VERSATILE -> DomainWeaponType.VERSATILE
-            WeaponType.TWO_HANDED_RANGED, WeaponType.RANGED -> DomainWeaponType.RANGED
-            else -> throw Exception("Bad Die value") // TODO: Implement proper Exception throwing with exception handlers
-        }
-    }
-}
