+++ domain/logic/src/main/kotlin/wasik/domain/logic/items/mapper/DomainActionMapper.kt
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Component
 import wasik.infrastructure.model.entity.ActionEntity
 
 @Component
-class DomainActionMapperImpl : ActionMapper {
-    override suspend fun mapToActionEntity(action: Action): ActionEntity = coroutineScope{
+class DomainActionMapper {
+    suspend fun mapToActionEntity(action: Action): ActionEntity = coroutineScope{
         ActionEntity(id = null, name = action.name, description = action.description)
     }
 }
