+++ domain/logic/src/main/kotlin/wasik/domain/logic/items/mapper/DomainPropertyMapper.kt
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Component
 import wasik.infrastructure.model.entity.PropertyEntity
 
 @Component
-class DomainPropertyMapperImpl : PropertyMapper {
-    override suspend fun mapToPropertyEntity(property: Property): PropertyEntity = coroutineScope {
+class DomainPropertyMapper {
+    suspend fun mapToPropertyEntity(property: Property): PropertyEntity = coroutineScope {
         PropertyEntity(id = null, name = property.name, description = property.description)
     }
 }
