--- domain/logic/src/main/kotlin/wasik/domain/logic/items/mapper/PropertyMapper.kt
@@ -1,8 +1,0 @@
-package wasik.domain.logic.items.mapper
-
-import domain.model.misc.Property
-import wasik.infrastructure.model.entity.PropertyEntity
-
-interface PropertyMapper {
-    suspend fun mapToPropertyEntity(property: Property): PropertyEntity
-}
